voice_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
elevenlabs==0.3.0
python-dotenv==1.0.1
boto3==1.34.29  # For S3 storage if needed 